Enable lula and dependent extensions in Windows build

+++ source/extensions/omni.isaac.motion_planning/plugins/MotionPolicy.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2020-2021, NVIDIA CORPORATION. All rights reserved.
+// Copyright (c) 2020-2022, NVIDIA CORPORATION. All rights reserved.
 //
 // NVIDIA CORPORATION and its licensors retain all intellectual property
 // and proprietary rights in and to this software, related documentation
@@ -176,7 +176,7 @@ void MotionPolicy::step(const float t, const float sourceDt)
             {
                 if (isnan(mState.pos()[idx]) == false)
                 {
-                    mDynamicControl->setDofPositionTarget(dof, mState.pos()[idx]);
+                    mDynamicControl->setDofPositionTarget(dof, static_cast<float>(mState.pos()[idx]));
                 }
                 else
                 {
